requirements.txt
Django==3.2
django-filter==2.4.0
drf-yasg==1.21.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
PyJWT==2.1.0
pytz==2021.1
python-dotenv==0.20.0
